backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pyngrok
transformers
geopy
nominatim
gunicorn